+++ pol_env/Tribes/src/core/game/Board.java
@@ -1042,6 +1042,7 @@ public class Board {
     public Types.BUILDING getBuildingAt(int x, int y){ return buildings[x][y]; }
     public void setUnits(int[][] u){ this.units = u; }
     public int getCityIdAt(int x, int y) { return tileCityId[x][y]; }
+    public void setCityIdAt(int x, int y, int cityId) { tileCityId[x][y] = cityId; }
     public int[] getCapitalIDs() {return capitalIDs;}
     boolean isNative() { return isNative; }
     public int getActorIDcounter() {